backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

jinja2==3.1.4
python-multipart==0.0.12

sqlalchemy==2.0.36

passlib==1.7.4
PyJWT==2.10.1

reportlab==4.2.5

bcrypt==5.0.0
